Major architecture change for the event callbacks
Event callbacks triggered from the backend caused a race condition when resizing the UI (trying to write to a window which was in the process of being reconstructed). Instead of using the callbacks to do a direct action on the UI, an event queue is added to buffer all events called from the backend and only making sure they can only be processed by the UI in a sequential way when the queue has event(s) buffered.

Same goes with the resizer being called previously by the signal interceptor in main(). Now only an atomic flag is raised leaving the UI to deal with the resizing call-stack on its own terms.